requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
plyer>=2.0.0
pyngrok>=5.1.0
python-dotenv>=0.19.0
pydantic>=1.8.2
